Refactor and use lock to avoid race conditions when shutting down
